extract the logic of version traversal into a new file
