+++ src/scope/component-ops/untag-component.ts
@@ -4,6 +4,7 @@ import logger from '../../logger/logger';
 import { Scope } from '..';
 import GeneralError from '../../error/general-error';
 import ComponentsList from '../../consumer/component/components-list';
+import { getAllVersionsObjects } from './traverse-versions';
 
 export type untagResult = { id: BitId; versions: string[]; component: ModelComponent };
 
@@ -48,7 +49,7 @@ export async function removeLocalVersion(
     });
   }
 
-  const allVersionsObjects = await component.getAllVersionsObjects(scope.objects);
+  const allVersionsObjects = await getAllVersionsObjects(component, scope.objects);
   scope.sources.removeComponentVersions(component, versionsToRemove, allVersionsObjects);
 
   return { id, versions: versionsToRemove, component };
